0_setup/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aioice==0.10.1
aiortc==1.13.0
aiosignal==1.3.2
aiounittest==1.3.0
annotated-types==0.7.0
anthropic==0.55.0
anyio==4.9.0
asyncer==0.0.7
attrs==25.3.0
av==14.4.0
azure-ai-agents==1.1.0b2
azure-ai-contentsafety==1.0.0
azure-ai-projects==1.0.0b11
azure-core==1.34.0
azure-identity==1.23.0
azure-storage-blob==12.25.1
Babel==2.9.1
backoff==2.2.1
bidict==0.23.1
botbuilder-core==4.17.0
botbuilder-dialogs==4.17.0
botbuilder-integration-aiohttp==4.17.0
botbuilder-schema==4.17.0
botframework-connector==4.17.0
botframework-streaming==4.17.0
certifi==2025.6.15
cffi==1.17.1
chainlit==2.5.5
chardet==5.2.0
charset-normalizer==3.4.2
chevron==0.14.0
click==8.2.1
cloudevents==1.12.0
colorama==0.4.6
cryptography==45.0.4
dataclasses-json==0.6.7
datedelta==1.4
defusedxml==0.7.1
Deprecated==1.2.18
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
emoji==1.7.0
fastapi==0.115.14
filelock==3.18.0
filetype==1.2.0
frozenlist==1.7.0
fsspec==2025.5.1
google-crc32c==1.7.1
googleapis-common-protos==1.70.0
grapheme==0.6.0
grpcio==1.73.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.33.1
idna==3.10
ifaddr==0.2.0
importlib_metadata==8.6.1
inflection==0.5.1
isodate==0.7.2
Jinja2==3.1.6
jiter==0.10.0
jsonpickle==1.4.2
jsonschema==4.24.0
jsonschema-path==0.3.4
jsonschema-specifications==2025.4.1
Lazify==0.4.0
lazy-object-proxy==1.11.0
literalai==0.1.201
MarkupSafe==3.0.2
marshmallow==3.26.1
mcp==1.10.0
monotonic==1.6
more-itertools==10.7.0
msal==1.32.3
msal-extensions==1.3.1
msrest==0.7.1
multidict==6.5.1
multipledispatch==1.0.0
mypy_extensions==1.1.0
nest-asyncio==1.6.0
numpy==2.3.1
oauthlib==3.3.1
openai==1.92.2
openapi-core==0.19.5
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.2
opentelemetry-api==1.31.1
opentelemetry-exporter-otlp==1.31.1
opentelemetry-exporter-otlp-proto-common==1.31.1
opentelemetry-exporter-otlp-proto-grpc==1.31.1
opentelemetry-exporter-otlp-proto-http==1.31.1
opentelemetry-instrumentation==0.52b1
opentelemetry-instrumentation-alephalpha==0.40.14
opentelemetry-instrumentation-anthropic==0.40.14
opentelemetry-instrumentation-bedrock==0.40.14
opentelemetry-instrumentation-chromadb==0.40.14
opentelemetry-instrumentation-cohere==0.40.14
opentelemetry-instrumentation-crewai==0.40.14
opentelemetry-instrumentation-google-generativeai==0.40.14
opentelemetry-instrumentation-groq==0.40.14
opentelemetry-instrumentation-haystack==0.40.14
opentelemetry-instrumentation-lancedb==0.40.14
opentelemetry-instrumentation-langchain==0.40.14
opentelemetry-instrumentation-llamaindex==0.40.14
opentelemetry-instrumentation-logging==0.52b1
opentelemetry-instrumentation-marqo==0.40.14
opentelemetry-instrumentation-mcp==0.40.14
opentelemetry-instrumentation-milvus==0.40.14
opentelemetry-instrumentation-mistralai==0.40.14
opentelemetry-instrumentation-ollama==0.40.14
opentelemetry-instrumentation-openai==0.40.14
opentelemetry-instrumentation-pinecone==0.40.14
opentelemetry-instrumentation-qdrant==0.40.14
opentelemetry-instrumentation-redis==0.52b1
opentelemetry-instrumentation-replicate==0.40.14
opentelemetry-instrumentation-requests==0.52b1
opentelemetry-instrumentation-sagemaker==0.40.14
opentelemetry-instrumentation-sqlalchemy==0.52b1
opentelemetry-instrumentation-threading==0.52b1
opentelemetry-instrumentation-together==0.40.14
opentelemetry-instrumentation-transformers==0.40.14
opentelemetry-instrumentation-urllib3==0.52b1
opentelemetry-instrumentation-vertexai==0.40.14
opentelemetry-instrumentation-watsonx==0.40.14
opentelemetry-instrumentation-weaviate==0.40.14
opentelemetry-proto==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
opentelemetry-semantic-conventions-ai==0.4.9
opentelemetry-util-http==0.52b1
packaging==25.0
parse==1.20.2
pathable==0.4.4
posthog==3.25.0
prance==25.4.8.0
propcache==0.3.2
protobuf==5.29.5
pybars4==0.9.13
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyee==13.0.0
PyJWT==2.10.1
pylibsrtp==0.12.0
PyMeta3==0.5.1
pyOpenSSL==25.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==4.12.2
python-multipart==0.0.18
python-socketio==5.13.0
pytz==2025.2
PyYAML==6.0.2
recognizers-text==1.0.2a2
recognizers-text-choice==1.0.2a2
recognizers-text-date-time==1.0.2a2
recognizers-text-number==1.0.2a2
recognizers-text-number-with-unit==1.0.2a2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rpds-py==0.25.1
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
scipy==1.16.0
semantic-kernel==1.34.0
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
sse-starlette==2.3.6
starlette==0.41.3
syncer==2.0.3
teams-ai==1.8.0
tenacity==9.1.2
tiktoken==0.9.0
tokenizers==0.21.2
tomli==2.2.1
tqdm==4.67.1
traceloop-sdk==0.40.14
types-PyYAML==6.0.12.20250516
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
uptrace==1.31.0
urllib3==2.5.0
uv==0.7.16
uvicorn==0.34.3
watchfiles==0.20.0
websockets==15.0.1
Werkzeug==3.1.1
wrapt==1.17.2
wsproto==1.2.0
yarl==1.20.1
zipp==3.23.0
bs4